+++ src/Phasor.cpp
@@ -1,6 +1,6 @@
 #include "Phasor.hpp"
 
-Phasor::Phasor(uint32_t freqHz, uint32_t sampleRateHz): phase(0), freqHz(freqHz), phaseIncStep(UINT32_MAX / (sampleRateHz / freqHz)) {}
+Phasor::Phasor(float freqHz, uint32_t sampleRateHz): phase(0), freqHz(freqHz), phaseIncStep(UINT32_MAX / (sampleRateHz / freqHz)) {}
 
 uint32_t Phasor::next()
 {
